LPS-114661 Rename UpgradeResourceActions to UpgradeResourceAction to match table name
