+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/upgrade/v1_0_2/UpgradeResourceAction.java
@@ -12,20 +12,21 @@
  * details.
  */
 
-package com.liferay.user.associated.data.web.internal.upgrade.v1_0_0;
+package com.liferay.portal.settings.web.internal.upgrade.v1_0_2;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
+import com.liferay.portal.settings.constants.PortalSettingsPortletKeys;
 
 /**
  * @author Tina Tian
  */
-public class UpgradeResourceActions extends UpgradeProcess {
+public class UpgradeResourceAction extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		runSQL(
-			"delete from ResourceAction where name = 'com_liferay_user_" +
-				"associated_data_web_portlet_UserAssociatedDataPortlet'");
+			"delete from ResourceAction where name = '" +
+				PortalSettingsPortletKeys.PORTAL_SETTINGS + "'");
 	}
 
 }
