+++ modules/apps/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/upgrade/UserAssociatedDataWebUpgrade.java
@@ -16,7 +16,7 @@ package com.liferay.user.associated.data.web.internal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.user.associated.data.web.internal.upgrade.v1_0_0.UpgradeResourceActions;
+import com.liferay.user.associated.data.web.internal.upgrade.v1_0_0.UpgradeResourceAction;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -30,7 +30,7 @@ public class UserAssociatedDataWebUpgrade implements UpgradeStepRegistrator {
 	public void register(Registry registry) {
 		registry.register("0.0.0", "1.0.0", new DummyUpgradeStep());
 
-		registry.register("0.0.1", "1.0.0", new UpgradeResourceActions());
+		registry.register("0.0.1", "1.0.0", new UpgradeResourceAction());
 	}
 
 }
