+++ modules/apps/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/upgrade/v1_0_0/UpgradeResourceAction.java
@@ -12,20 +12,20 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_3_x;
+package com.liferay.user.associated.data.web.internal.upgrade.v1_0_0;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 
 /**
- * @author Dante Wang
+ * @author Tina Tian
  */
-public class UpgradeResourceActions extends UpgradeProcess {
+public class UpgradeResourceAction extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		runSQL(
-			"delete from ResourceAction where name in ('136', '150', '151', " +
-				"'152', '153', '157', '158')");
+			"delete from ResourceAction where name = 'com_liferay_user_" +
+				"associated_data_web_portlet_UserAssociatedDataPortlet'");
 	}
 
 }
