+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_6.java
@@ -17,7 +17,7 @@ package com.liferay.portal.upgrade;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
 import com.liferay.portal.upgrade.v7_0_6.UpgradeRepository;
-import com.liferay.portal.upgrade.v7_0_6.UpgradeResourceActions;
+import com.liferay.portal.upgrade.v7_0_6.UpgradeResourceAction;
 import com.liferay.portal.upgrade.v7_0_6.UpgradeThemeId;
 
 /**
@@ -33,7 +33,7 @@ public class UpgradeProcess_7_0_6 extends UpgradeProcess {
 	@Override
 	protected void doUpgrade() throws Exception {
 		upgrade(new UpgradeRepository());
-		upgrade(new UpgradeResourceActions());
+		upgrade(new UpgradeResourceAction());
 		upgrade(new UpgradeThemeId());
 
 		clearIndexesCache();
